+++ docs/source/en/using-diffusers/loading_adapters.md
@@ -485,7 +485,7 @@ image.save("sdxl_t2i.png")
   </div>
 </div>
 
-You can use the IP-Adapter face models to apply specific faces to your images.  It is an effective way to maintain consistent characters in your image generations.
+You can use the IP-Adapter face model to apply specific faces to your images.  It is an effective way to maintain consistent characters in your image generations.
 Weights are loaded with the same method used for the other IP-Adapters.  
 
 ```python
@@ -495,7 +495,7 @@ pipeline.load_ip_adapter("h94/IP-Adapter", subfolder="models", weight_name="ip-a
 
 <Tip>
 
-It is recommended to use `DDIMScheduler` and `EulerDiscreteScheduler` for face models. 
+It is recommended to use `DDIMScheduler` and `EulerDiscreteScheduler` for face model. 
 
 
 </Tip>
@@ -548,58 +548,6 @@ image = pipeline(
     <figcaption class="mt-2 text-center text-sm text-gray-500">output image</figcaption>
   </div>
 </div>
-
-IP Adapter FaceID is an experimental IP Adapter model that uses image embeddings generated by `insightface`, so no image encoder needs to be loaded.
-You need to install `insightface` and all its requirements to use this model.
-You must pass the image embedding tensor as `image_embeds` to the StableDiffusionPipeline instead of `ip_adapter_image`.
-
-```
-import cv2
-from insightface.app import FaceAnalysis
-import numpy as np
-from PIL import Image
-import torch
-from diffusers import StableDiffusionPipeline, DDIMScheduler
-from diffusers.utils import load_image
-
-noise_scheduler = DDIMScheduler(
-    num_train_timesteps=1000,
-    beta_start=0.00085,
-    beta_end=0.012,
-    beta_schedule="scaled_linear",
-    clip_sample=False,
-    set_alpha_to_one=False,
-    steps_offset=1
-)
-
-pipeline = StableDiffusionPipeline.from_pretrained(
-    "SG161222/Realistic_Vision_V4.0_noVAE",
-    torch_dtype=torch.float16,
-    scheduler=noise_scheduler
-).to("cuda")
-
-generator = torch.Generator(device="cpu").manual_seed(42)
-image = load_image("https://huggingface.co/datasets/YiYiXu/testing-images/resolve/main/ai_face2.png")
-
-# Extract image embeddings
-app = FaceAnalysis(name="buffalo_l", providers=['CUDAExecutionProvider', 'CPUExecutionProvider'])
-app.prepare(ctx_id=0, det_size=(640, 640))
-
-image = cv2.cvtColor(np.asarray(image), cv2.COLOR_BGR2RGB)
-faces = app.get(image)
-image = torch.from_numpy(faces[0].normed_embedding).unsqueeze(0)
-
-# Load IP Adapter weights and run inference
-pipeline.load_ip_adapter("h94/IP-Adapter-FaceID", subfolder=None, weight_name="ip-adapter-faceid_sd15.bin")
-pipeline.set_ip_adapter_scale(0.7)
-images = pipeline(
-  prompt="A photo of a girl wearing a black dress, holding red roses in hand, upper body, behind is the Eiffel Tower",
-  image_embeds=image,
-  negative_prompt="monochrome, lowres, bad anatomy, worst quality, low quality", 
-  num_inference_steps=20, num_images_per_prompt=1, width=512, height=704, 
-  generator=generator
-).images[0]
-```
 
 ### LCM-Lora
