+++ src/diffusers/loaders/unet.py
@@ -707,20 +707,13 @@ class UNet2DConditionLoadersMixin:
                 diffusers_name = key.replace("proj", "image_embeds")
                 updated_state_dict[diffusers_name] = value
 
-        elif "proj.0.weight" in state_dict:
+        elif "proj.3.weight" in state_dict:
             # IP-Adapter Full
-            clip_embeddings_dim_in = state_dict["proj.0.weight"].shape[1]
-            clip_embeddings_dim_out = state_dict["proj.0.weight"].shape[0]
-            multiplier = clip_embeddings_dim_out // clip_embeddings_dim_in
-            norm_layer = "proj.3.weight" if "proj.3.weight" in state_dict else "norm.weight"
-            cross_attention_dim = state_dict[norm_layer].shape[0]
-            num_tokens = state_dict["proj.2.weight"].shape[0] // cross_attention_dim
+            clip_embeddings_dim = state_dict["proj.0.weight"].shape[0]
+            cross_attention_dim = state_dict["proj.3.weight"].shape[0]
 
             image_projection = IPAdapterFullImageProjection(
-                cross_attention_dim=cross_attention_dim,
-                image_embed_dim=clip_embeddings_dim_in,
-                mult=multiplier,
-                num_tokens=num_tokens,
+                cross_attention_dim=cross_attention_dim, image_embed_dim=clip_embeddings_dim
             )
 
             for key, value in state_dict.items():
@@ -774,23 +767,14 @@ class UNet2DConditionLoadersMixin:
             AttnProcessor2_0,
             IPAdapterAttnProcessor,
             IPAdapterAttnProcessor2_0,
-            LoRAIPAdapterAttnProcessor,
-            LoRAIPAdapterAttnProcessor2_0,
         )
-        from ..models.lora import LoRALinearLayer
 
-        use_lora = False
         if "proj.weight" in state_dict["image_proj"]:
             # IP-Adapter
             num_image_text_embeds = 4
-        elif "proj.0.weight" in state_dict["image_proj"]:
+        elif "proj.3.weight" in state_dict["image_proj"]:
             # IP-Adapter Full Face
             num_image_text_embeds = 257  # 256 CLIP tokens + 1 CLS token
-            for k in state_dict["ip_adapter"].keys():
-                if "lora" in k:
-                    num_image_text_embeds = 4
-                    use_lora = True
-                    break
         else:
             # IP-Adapter Plus
             num_image_text_embeds = state_dict["image_proj"]["latents"].shape[1]
@@ -801,7 +785,7 @@ class UNet2DConditionLoadersMixin:
 
         # set ip-adapter cross-attention processors & load state_dict
         attn_procs = {}
-        key_id = 1 if not use_lora else 0
+        key_id = 1
         for name in self.attn_processors.keys():
             cross_attention_dim = None if name.endswith("attn1.processor") else self.config.cross_attention_dim
             if name.startswith("mid_block"):
@@ -817,95 +801,23 @@ class UNet2DConditionLoadersMixin:
                     AttnProcessor2_0 if hasattr(F, "scaled_dot_product_attention") else AttnProcessor
                 )
                 attn_procs[name] = attn_processor_class()
-                if use_lora:
-                    rank = state_dict["ip_adapter"][f"{key_id}.to_q_lora.down.weight"].shape[0]
-                    attn_module = self
-                    for n in name.split(".")[:-1]:
-                        attn_module = getattr(attn_module, n)
-                    # Set the `lora_layer` attribute of the attention-related matrices.
-                    attn_module.to_q.set_lora_layer(
-                        LoRALinearLayer(
-                            in_features=attn_module.to_q.in_features,
-                            out_features=attn_module.to_q.out_features,
-                            rank=rank,
-                        )
-                    )
-                    attn_module.to_k.set_lora_layer(
-                        LoRALinearLayer(
-                            in_features=attn_module.to_k.in_features,
-                            out_features=attn_module.to_k.out_features,
-                            rank=rank,
-                        )
-                    )
-                    attn_module.to_v.set_lora_layer(
-                        LoRALinearLayer(
-                            in_features=attn_module.to_v.in_features,
-                            out_features=attn_module.to_v.out_features,
-                            rank=rank,
-                        )
-                    )
-                    attn_module.to_out[0].set_lora_layer(
-                        LoRALinearLayer(
-                            in_features=attn_module.to_out[0].in_features,
-                            out_features=attn_module.to_out[0].out_features,
-                            rank=rank,
-                        )
-                    )
-
-                    value_dict = {}
-                    for k, module in attn_module.named_children():
-                        index = "."
-                        if not hasattr(module, "set_lora_layer"):
-                            index = ".0."
-                            module = module[0]
-                        lora_layer = getattr(module, "lora_layer")
-                        for lora_name, w in lora_layer.state_dict().items():
-                            value_dict.update(
-                                {
-                                    f"{k}{index}lora_layer.{lora_name}": state_dict["ip_adapter"][
-                                        f"{key_id}.{k}_lora.{lora_name}"
-                                    ]
-                                }
-                            )
-
-                    attn_module.load_state_dict(value_dict, strict=False)
-                    attn_module.to(dtype=self.dtype, device=self.device)
-                    key_id += 1
             else:
-                if use_lora:
-                    rank = state_dict["ip_adapter"][f"{key_id}.to_q_lora.down.weight"].shape[0]
-                    attn_processor_class = (
-                        LoRAIPAdapterAttnProcessor2_0
-                        if hasattr(F, "scaled_dot_product_attention")
-                        else LoRAIPAdapterAttnProcessor
-                    )
-                    attn_procs[name] = attn_processor_class(
-                        hidden_size=hidden_size,
-                        cross_attention_dim=cross_attention_dim,
-                        scale=1.0,
-                        rank=rank,
-                        num_tokens=num_image_text_embeds,
-                    ).to(dtype=self.dtype, device=self.device)
-
-                else:
-                    attn_processor_class = (
-                        IPAdapterAttnProcessor2_0
-                        if hasattr(F, "scaled_dot_product_attention")
-                        else IPAdapterAttnProcessor
-                    )
-                    attn_procs[name] = attn_processor_class(
-                        hidden_size=hidden_size,
-                        cross_attention_dim=cross_attention_dim,
-                        scale=1.0,
-                        num_tokens=num_image_text_embeds,
-                    ).to(dtype=self.dtype, device=self.device)
+                attn_processor_class = (
+                    IPAdapterAttnProcessor2_0 if hasattr(F, "scaled_dot_product_attention") else IPAdapterAttnProcessor
+                )
+                attn_procs[name] = attn_processor_class(
+                    hidden_size=hidden_size,
+                    cross_attention_dim=cross_attention_dim,
+                    scale=1.0,
+                    num_tokens=num_image_text_embeds,
+                ).to(dtype=self.dtype, device=self.device)
 
                 value_dict = {}
                 for k, w in attn_procs[name].state_dict().items():
                     value_dict.update({f"{k}": state_dict["ip_adapter"][f"{key_id}.{k}"]})
 
                 attn_procs[name].load_state_dict(value_dict)
-                key_id += 2 if not use_lora else 1
+                key_id += 2
 
         self.set_attn_processor(attn_procs)
